app/pyproject.toml
[tool.poetry]
name = "anki-chinese"
version = "0.1.0"
description = ""
authors = ["Rocket <rocketnova@users.noreply.github.com>"]
license = "GPL-3.0-or-later"
repository = "https://github.com/rocketnova/anki_chinese"
packages = [{include = "anki_chinese"}]


[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
dragonmapper = "^0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
test = "anki_chinese.test:main"
csv = "anki_chinese.csv:main"

